requirements.txt
flask
requests
beautifulsoup4
youtube-transcript-api
openai
pyyaml
pytest
